+++ src/camera.rs
@@ -4,36 +4,36 @@ use self::cgmath::InnerSpace;
 use self::cgmath::Vector3;
 use self::rand::prelude::*;
 use ray::Ray;
-use std::f32;
+use std::f64;
 
-fn random_in_unit_sphere() -> Vector3<f32> {
-    Vector3::new(random::<f32>(), random::<f32>(), random::<f32>()).normalize()
+fn random_in_unit_sphere() -> Vector3<f64> {
+    Vector3::new(random::<f64>(), random::<f64>(), random::<f64>()).normalize()
 }
 
 pub struct Camera {
-    pub lower_left: Vector3<f32>,
-    pub horizontal: Vector3<f32>,
-    pub vertical: Vector3<f32>,
-    pub origin: Vector3<f32>,
-    pub u: Vector3<f32>,
-    pub v: Vector3<f32>,
-    pub w: Vector3<f32>,
-    pub lens_radius: f32,
+    pub lower_left: Vector3<f64>,
+    pub horizontal: Vector3<f64>,
+    pub vertical: Vector3<f64>,
+    pub origin: Vector3<f64>,
+    pub u: Vector3<f64>,
+    pub v: Vector3<f64>,
+    pub w: Vector3<f64>,
+    pub lens_radius: f64,
 }
 
 impl Camera {
     pub fn new(
-        look_from: Vector3<f32>,
-        look_at: Vector3<f32>,
-        vup: Vector3<f32>,
-        vfov: f32,
-        aspect: f32,
-        aperture: f32,
-        focus_dist: f32,
+        look_from: Vector3<f64>,
+        look_at: Vector3<f64>,
+        vup: Vector3<f64>,
+        vfov: f64,
+        aspect: f64,
+        aperture: f64,
+        focus_dist: f64,
     ) -> Camera {
         let (u, v, w);
         let lens_radius = aperture / 2.0;
-        let theta = vfov * f32::consts::PI / 180.0;
+        let theta = vfov * f64::consts::PI / 180.0;
         let half_height = (theta / 2.0).tan();
         let half_width = aspect * half_height;
         let origin = look_from;
@@ -58,9 +58,9 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, u: f32, v: f32) -> Ray {
+    pub fn get_ray(&self, u: f64, v: f64) -> Ray {
         let rd = self.lens_radius * random_in_unit_sphere();
-        let offset: Vector3<f32> = self.u * rd.x + self.v * rd.y;
+        let offset: Vector3<f64> = self.u * rd.x + self.v * rd.y;
         Ray::new(
             &(self.origin + offset),
             self.lower_left + u * self.horizontal + v * self.vertical - self.origin - offset,
